seperated gaetteer into its own file but it's not working - Scottish-Natural-Heritage/Licensing#1212
