+++ src/controllers/choose-address.js
@@ -1,51 +1,10 @@
 import {ReturnState} from './_base.js';
-import axios from 'axios';
-import config from '../config.js';
-
-/**
- * Find addresses by postcode.
- *
- * @param {number} uprn The UPRN to find addresses by.
- */
-const findFullAddressesByUprn = async (uprn) => {
-  let apiResponse;
-
-  try {
-    // Lookup the postcode in our Gazetteer API.
-    apiResponse = await axios.get(config.gazetteerApiEndpoint, {
-      params: {
-        uprn,
-        fieldset: 'all'
-      },
-      headers: {
-        Authorization: `Bearer ${config.gazetteerApiKey}`
-      },
-      timeout: 10_000
-    });
-  } catch (error) {
-    console.log(error);
-  }
-
-  // Grab just the json payload.
-  const apiData = apiResponse.data;
-
-  // A single string in the array rather than an array of objects indicates an
-  // error where no addresses have been found.
-  if (apiData.metadata.count === 0 || (apiData.results.length === 1 && typeof apiData.results[0] === 'string')) {
-    throw new Error('No matching addresses found.');
-  }
-
-  // Treat the json blob as a typed response.
-  const gazetteerResponse = apiData;
-
-  // Dig out the right array from the returned json blob.
-  return gazetteerResponse.results[0].address;
-};
+import Gazetteer from '../utils/gazetteer.js';
 
 const chooseAddressController = async (request) => {
   try {
     // Get full address uprn
-    const gazetteerAddresses = await findFullAddressesByUprn(request.session.uprn ?? 0);
+    const gazetteerAddresses = await Gazetteer.findFullAddressesByUprn(request.session.uprn ?? 0);
 
     // Build up address line 1
     const subBuildingName = gazetteerAddresses[0].sub_building_name
